prototype/AIML_module/botenv_requirements.txt
Package                   Version
------------------------- -------------------
aiml                      0.9.2
butterfly                 3.2.5
cachetools                4.2.2
certifi                   2021.5.30
cffi                      1.14.6
charset-normalizer        2.0.4
click                     8.0.1
cryptography              3.4.7
Flask                     2.0.1
google-api-core           1.31.0
google-auth               1.32.1
google-cloud-speech       2.5.0
google-cloud-texttospeech 2.5.0
googleapis-common-protos  1.53.0
grpcio                    1.38.1
idna                      3.2
itsdangerous              2.0.1
Jinja2                    3.0.1
libcst                    0.3.19
MarkupSafe                2.0.1
mypy-extensions           0.4.3
packaging                 21.0
pip                       21.0.1
proto-plus                1.19.0
protobuf                  3.17.3
pyasn1                    0.4.8
pyasn1-modules            0.2.8
PyAudio                   0.2.11
pycparser                 2.20
pygame                    2.0.1
pyOpenSSL                 20.0.1
pyparsing                 2.4.7
python-aiml               0.9.3
pytz                      2021.1
requests                  2.26.0
rsa                       4.7.2
setuptools                52.0.0.post20210125
six                       1.16.0
SpeechRecognition         3.8.1
tornado                   6.1
typing-extensions         3.10.0.0
typing-inspect            0.7.1
urllib3                   1.26.6
webrtcvad                 2.0.10
Werkzeug                  2.0.1
wheel                     0.37.0
